+++ src/main/java/de/unistuttgart/ims/coref/annotator/action/NewEntityAction.java
@@ -14,7 +14,7 @@ import de.unistuttgart.ims.coref.annotator.DocumentWindow;
 import de.unistuttgart.ims.coref.annotator.Strings;
 import de.unistuttgart.ims.coref.annotator.document.op.AddMentionsToNewEntity;
 
-public class NewEntityAction extends TargetedIkonAction<DocumentWindow> {
+public class NewEntityAction extends TargetedOperationIkonAction<DocumentWindow> {
 
 	private static final long serialVersionUID = 1L;
 
@@ -22,8 +22,9 @@ public class NewEntityAction extends TargetedIkonAction<DocumentWindow> {
 		super(dw, Strings.ACTION_NEW, MaterialDesign.MDI_ACCOUNT_PLUS);
 		putValue(Action.SHORT_DESCRIPTION, Annotator.getString(Strings.ACTION_NEW_TOOLTIP));
 		putValue(Action.ACCELERATOR_KEY,
-				KeyStroke.getKeyStroke(KeyEvent.VK_N, Toolkit.getDefaultToolkit().getMenuShortcutKeyMask()));
+				KeyStroke.getKeyStroke(KeyEvent.VK_N, Toolkit.getDefaultToolkit().getMenuShortcutKeyMaskEx()));
 
+		operationClass = AddMentionsToNewEntity.class;
 	}
 
 	@Override