+++ src/main/java/de/unistuttgart/ims/coref/annotator/action/OperationAction.java
@@ -1,0 +1,8 @@
+package de.unistuttgart.ims.coref.annotator.action;
+
+import de.unistuttgart.ims.coref.annotator.document.op.Operation;
+
+public interface OperationAction {
+	Class<? extends Operation> getOperationClass();
+
+}