+++ src/main/java/de/unistuttgart/ims/coref/annotator/action/ViewFontSizeDecreaseAction.java
@@ -20,7 +20,7 @@ public class ViewFontSizeDecreaseAction extends DocumentWindowAction {
 	public ViewFontSizeDecreaseAction(DocumentWindow documentWindow) {
 		super(documentWindow, Strings.ACTION_VIEW_DECREASE_FONT_SIZE, MaterialDesign.MDI_FORMAT_SIZE);
 		putValue(Action.ACCELERATOR_KEY,
-				KeyStroke.getKeyStroke(KeyEvent.VK_MINUS, Toolkit.getDefaultToolkit().getMenuShortcutKeyMask()));
+				KeyStroke.getKeyStroke(KeyEvent.VK_MINUS, Toolkit.getDefaultToolkit().getMenuShortcutKeyMaskEx()));
 
 	}