+++ src/main/java/de/unistuttgart/ims/coref/annotator/worker/DocumentModelLoader.java
@@ -10,11 +10,13 @@ import org.apache.uima.jcas.JCas;
 
 import de.unistuttgart.ims.coref.annotator.Annotator;
 import de.unistuttgart.ims.coref.annotator.document.DocumentModel;
+import de.unistuttgart.ims.coref.annotator.profile.Profile;
 
 public class DocumentModelLoader extends SwingWorker<DocumentModel, Integer> {
 
 	Consumer<DocumentModel> consumer = null;
 	JCas jcas;
+	Profile profile;
 
 	public DocumentModelLoader(Consumer<DocumentModel> consumer, JCas jcas) {
 		this.consumer = consumer;
@@ -24,15 +26,16 @@ public class DocumentModelLoader extends SwingWorker<DocumentModel, Integer> {
 	protected DocumentModel load(Preferences preferences) {
 		Annotator.logger.debug("Starting loading of coreference model");
 		DocumentModel documentModel = new DocumentModel(jcas, preferences);
-
 		documentModel.initialize();
-
+		if (profile != null)
+			documentModel.loadProfile(profile);
 		return documentModel;
 	}
 
 	@Override
 	protected DocumentModel doInBackground() throws Exception {
-		return load(Annotator.app.getPreferences());
+		DocumentModel documentModel = load(Annotator.app.getPreferences());
+		return documentModel;
 	}
 
 	@Override
@@ -42,6 +45,14 @@ public class DocumentModelLoader extends SwingWorker<DocumentModel, Integer> {
 		} catch (InterruptedException | ExecutionException e) {
 			Annotator.logger.catching(e);
 		}
+	}
+
+	public Profile getProfile() {
+		return profile;
+	}
+
+	public void setProfile(Profile profile) {
+		this.profile = profile;
 	}
 
 }